move `maybe_slow_query` and `failed_query` to `logger` crate.
