+++ proxy/src/lib.rs
@@ -32,7 +32,6 @@ pub mod instance;
 pub mod limiter;
 mod metrics;
 pub mod opentsdb;
-mod query_log;
 mod read;
 pub mod schema_config_provider;
 mod util;
